lib.sh: trap: handle a gotcha with `set -e` and nested trap scopes
When `set -e` is used with globaltraps and ltraps, if an error occurs
in an inner scope, only the inner scope runs before shell terminates.

This happens because errors are not converted into returns. Thus only
the globaltraps handler is run on an error, and it picks up just the
closest (inner) scope.

Fix this by making the globaltraps handler run through _all_ trap scopes
by iteratively unsetting __traps and re-running if it is still set.

Note that any of this won't happen if only ltraps scopes are active.
Thus, if `set -e` is used, globaltraps must be active for traps to run.