lib.sh: getopt: WTF moment: array element unset + nullglob
$ shopt -s nullglob
$ declare -A ARR=([foo]=FOO [bar]=BAR)
$ declare -p ARR
declare -A ARR=([foo]="FOO" [bar]="BAR" )
$ unset ARR[foo]
$ declare -p ARR
declare -A ARR=([foo]="FOO" [bar]="BAR" )

This happens because `ARR[foo]` is treated as an expansion
and predictably matches no files, thus expanding to an empty string.
I really need to fix this properly, because if `ARRf` or `ARRo` happen
to exist in cwd, this will expand to a bogus value even without nullglob.