lib.sh: trap: WTF moment: use a separate flag variable for each scope
In bash, an empty array is treated as "unset" for the purposes of the
${var+text} alternative expansion. Thus, if __traps array happens to be
empty in a particular scope because no traps have been registered,
testing ${__traps+set} will erroneously indicate absence of a scope
(thus any shadowed scopes will not be unwound).

Fix this by declaring a separate scalar flag variable in each scope
which will always predictably expand to a non-empty value.
(Also, set a unique value equal to the call stack depth and test the
value itself for ease of `set -x` debugging.)