lib.sh: trap: WTF moment: work around new unset semantics
In bash, executing an `unset` on a local variable behaves differently
depending on whether this is a local variable _in the current scope_
or a local variable _in a parent scope_ (dynamically visible due to
normal bash scoping rules). In the former case, the variable is
replaced with a sort of a "whiteout", preventing a shadowed variable
from the parent scope (if one exists) from being visible in this scope.

However, in the latter case no such whiteout is created and a shadowed
variable from the "grandparent" scope will become visible.

Since we want the latter behavior to be able to process all parent
scopes with traps in turn by repeatedly unsetting $__traps, replace
`unset` with a wrapper function `__dynamic_unset()` (which simply
calls normal `unset` on its arguments), causing our unsets to operate
effectively on the parent scope (upholding our desired semantics).