lib.sh: argv0: also set `$LIB_ARGV` roughly equivalent to `$0 $*`
Add an additional variable `$LIB_ARGV` that is roughly equivalent to
`$0 $*` (useful for logging or debugging), but doing the right thing
(i.e. using `$LIB_ARGV0`, quoting arguments (but not `$0`, unless we
can figure out a way to quote-escape an argument only if needed) and
omitting the space before `$*` if there are no arguments.